Improve error messages in dashboards_controller

+++ api/lib/mno_enterprise/concerns/controllers/jpi/v1/impac/dashboards_controller.rb
@@ -19,42 +19,46 @@ module MnoEnterprise::Concerns::Controllers::Jpi::V1::Impac::DashboardsControlle
   end
 
   # GET /mnoe/jpi/v1/impac/dashboards/1
+  #   -> GET /api/mnoe/v1/users/1/dashboards
   def show
     dashboard
-    render json: { errors: "Dashboard id #{params[:id]} doesn't exist" }, status: :not_found unless @dashboard
+    render_not_found('dashboard') unless @dashboard
   end
 
   # POST /mnoe/jpi/v1/impac/dashboards
-  #  -> POST /api/mnoe/v1/users/282/dashboards
+  #   -> POST /api/mnoe/v1/users/1/dashboards
   def create
     if @dashboard = dashboards.create(dashboard_create_params)
       # authorize! :create, @dashboard
       MnoEnterprise::EventLogger.info('dashboard_create', current_user.id, 'Dashboard Creation', nil, @dashboard)
       render 'show'
     else
-      render json: @dashboard.errors, status: :bad_request
+      render_bad_request('create dashboard', @dashboard.errors)
     end
   end
 
   # PUT /mnoe/jpi/v1/impac/dashboards/1
   def update
+    return render_not_found('dashboard') unless dashboard
+
     if dashboard.update(dashboard_update_params)
-      # dashboard.assign_attributes(attrs)
       # authorize! :update, dashboard
       render 'show'
     else
-      render json: @dashboard.errors, status: :bad_request
+      render_bad_request('update dashboard', dashboard.errors)
     end
   end
 
   # DELETE /mnoe/jpi/v1/impac/dashboards/1
   def destroy
-    # authorize! :destroy, @dashboard
+    return render_not_found('dashboard') unless dashboard
+
     if dashboard.destroy
+      # authorize! :destroy, @dashboard
       MnoEnterprise::EventLogger.info('dashboard_delete', current_user.id, 'Dashboard Deletion', nil, dashboard)
       head status: :ok
     else
-      render json: 'Unable to destroy dashboard', status: :bad_request
+      render_bad_request('destroy dashboard', 'Unable to destroy dashboard')
     end
   end
 
@@ -76,7 +80,7 @@ module MnoEnterprise::Concerns::Controllers::Jpi::V1::Impac::DashboardsControlle
   # for the Her "meta_data" issue.
   def dashboard_params
     params.require(:dashboard).permit(*whitelisted_params).tap do |whitelisted|
-      whitelisted[:settings] = params[:dashboard][:metadata]
+      whitelisted[:settings] = params[:dashboard][:metadata] || {}
     end
     .except(:metadata)
   end
